in the home page Hero section at the top of the page, I would like for you to add this in the center  in white, easily readable font.
"All rates and fees stated and discussed are for my time only, with the purpose and intent of serving as your personal enrichment guide. I accomplish this using the knowledge I have acquired as a self-taught alternative sex-positive therapist; aiding in the improvement of sexual and psychological health, happiness, competence, and emotional and spiritual well-being. As such, the only services I provide are consulting, counseling, coaching, and therapeutic introspection training. No quid pro quo exists in regards to my time or services."