add a fancy, page divider/section divider UI visual between the Service Packes floating cards and 'Escort Service Activities'
also, in the footer section. I would like quick links of the links found in my navbar, in the center of the footer, it should read: © 2025 Nikki By Night - TLee Apps. All rights reserved.
but make that slightly smaller,
move the 'Legal Disclaimer to the far right, replacing the 3 social media icons.